requirements.txt
pymongo==4.6.2 
pandas==2.1.4   
apache-airflow-providers-mongo==3.3.0  

requests==2.31.0
python-dotenv==1.0.0
pika==1.3.2